requirements.txt
flask
flask-cors
torch
torchvision
pillow
gunicorn
gdown
waitress
requests